% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt_rp.R
\name{bpt_rp}
\alias{bpt_rp}
\title{A function to fit BPT renewal model}
\usage{
bpt_rp(data, t, m, y)
}
\arguments{
\item{data}{input inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{m}{the number of iterations in nlm}

\item{y}{user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting BPT renewal model
\describe{
\item{par1}{Estimated parameter (mu) of the BPT model}
\item{par2}{Estimated parameter (alpha) of the BPT model}
\item{logL}{Negative log-likelihood}
\item{AIC}{Akaike information criterion (AIC)}
\item{BIC}{Bayesian information criterion (BIC)}
\item{mu_hat}{Estimated mean}
\item{pr_hat}{Estimated (logit) probabilities}
\item{haz_hat}{Estimated (log) hazard rates}
}
}
\description{
A function to fit BPT renewal model
}
\examples{
set.seed(42)
data <-  rgamma(30,3,0.01)

# set some parameters
m <- 10  # number of iterations for MLE optimization
t <- seq(100, 200, by=10)  # time intervals
y <- 304  # cut-off year for estimating probablity

# fit BPT renewal model
fit <- marp::bpt_rp(data, t, m, y)
fit
summary(fit)

}
