% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{coef.marp_model_fit}
\alias{coef.marp_model_fit}
\title{Extract parameters from a model-specific fit}
\usage{
\method{coef}{marp_model_fit}(object, ...)
}
\arguments{
\item{object}{A `marp_model_fit` object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A named numeric vector of fitted distribution parameters.
}
\description{
Extract parameters from a model-specific fit
}
