% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dllog.R
\name{dllog}
\alias{dllog}
\title{Density function of Log-Logistics model}
\usage{
dllog(x, shape = 1, scale = 1, log = FALSE)
}
\arguments{
\item{x}{Numeric vector of positive quantiles.}

\item{shape}{Positive shape parameter of the log-logistic distribution.}

\item{scale}{Positive scale parameter of the log-logistic distribution.}

\item{log}{Logical; if `TRUE`, return log-densities.}
}
\value{
A numeric vector of densities, or log-densities when `log = TRUE`.
}
\description{
Density function of Log-Logistics model
}
\examples{
x <- as.numeric(c(350., 450., 227., 352., 654.))
# set paramters
shape <- 5
scale <- 3
log <- FALSE
result_1 <- marp::dllog(x, shape, scale, log)

# alternatively, set log == TRUE
log <- TRUE
result_2 <- marp::dllog(x, shape, scale, log)

}
