% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{logLik.marp_model_fit}
\alias{logLik.marp_model_fit}
\title{Extract the log-likelihood from a model-specific fit}
\usage{
\method{logLik}{marp_model_fit}(object, ...)
}
\arguments{
\item{object}{A `marp_model_fit` object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An object of class `logLik`.
}
\description{
Extract the log-likelihood from a model-specific fit
}
