% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognorm_rp.R
\name{lognorm_rp}
\alias{lognorm_rp}
\title{A function to fit Log-Normal renewal model}
\usage{
lognorm_rp(data, t, y)
}
\arguments{
\item{data}{A numeric vector of positive inter-event times.}

\item{t}{A numeric vector of time points at which log-hazards are evaluated.}

\item{y}{A time point at which the logit-transformed cumulative event
probability is evaluated.}
}
\value{
An object of class `marp_model_fit` retaining the following eight
  named list components:
\describe{
\item{par1}{Estimated mean (on the log scale) of the Log-Normal model}
\item{par2}{Estimated standard deviation (on the log scale)of the Log-Normal model}
\item{logL}{Maximized log-likelihood}
\item{AIC}{Akaike information criterion (AIC)}
\item{BIC}{Bayesian information criterion (BIC)}
\item{mu_hat}{Estimated mean inter-event time}
\item{pr_hat}{Logit-transformed cumulative event probability at `y`}
\item{haz_hat}{Log-hazard values at `t`}
}
}
\description{
A function to fit Log-Normal renewal model
}
\examples{
set.seed(42)
data <-  rgamma(100,3,0.01)

# set some parameters
t = seq(100, 200, by=10)  # time intervals
y = 304  # cut-off year for estimating probablity

# fit Log-Normal renewal model
fit <- marp::lognorm_rp(data, t, y)
fit
summary(fit)

}
