% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{nobs.marp_fit}
\alias{nobs.marp_fit}
\title{Number of observations in a model-averaged fit}
\usage{
\method{nobs}{marp_fit}(object, ...)
}
\arguments{
\item{object}{A `marp_fit` object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The number of observations used for fitting.
}
\description{
Number of observations in a model-averaged fit
}
