% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pllog.R
\name{pllog}
\alias{pllog}
\title{Probability function of Log-Logistics model}
\usage{
pllog(q, shape = 1, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Numeric vector of positive quantiles.}

\item{shape}{Positive shape parameter of the log-logistic distribution.}

\item{scale}{Positive scale parameter of the log-logistic distribution.}

\item{lower.tail}{Logical; if `TRUE`, return lower-tail probabilities,
otherwise return upper-tail probabilities.}

\item{log.p}{Logical; if `TRUE`, return probabilities on the log scale.}
}
\value{
A numeric vector of probabilities, or log-probabilities when
  `log.p = TRUE`.
}
\description{
Probability function of Log-Logistics model
}
\examples{
q <- c(1, 2, 3, 4)
# set paramters
shape <- 5
scale <- 3
result_1 <- marp::pllog(q, shape, scale)

# alternatively, return log-probabilities
result_2 <- marp::pllog(q, shape, scale, log.p = TRUE)

}
