% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{print.marp_fit}
\alias{print.marp_fit}
\title{Print a model-averaged renewal-process fit}
\usage{
\method{print}{marp_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by [marp()].}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a model-averaged renewal-process fit
}
