% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{print.marp_model_fit}
\alias{print.marp_model_fit}
\title{Print a model-specific renewal-process fit}
\usage{
\method{print}{marp_model_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by one of the model-specific fitting functions.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a model-specific renewal-process fit
}
