% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{print.summary_marp_fit}
\alias{print.summary_marp_fit}
\title{Print a summary of a model-averaged renewal-process fit}
\usage{
\method{print}{summary_marp_fit}(x, ...)
}
\arguments{
\item{x}{A `summary_marp_fit` object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a summary of a model-averaged renewal-process fit
}
