% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{summary.marp_fit}
\alias{summary.marp_fit}
\title{Summarise a model-averaged renewal-process fit}
\usage{
\method{summary}{marp_fit}(object, ...)
}
\arguments{
\item{object}{An object returned by [marp()].}

\item{...}{Additional arguments, currently unused.}
}
\value{
A structured object of class `summary_marp_fit`.
}
\description{
Summarise a model-averaged renewal-process fit
}
