% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_methods.R
\name{summary.marp_model_fit}
\alias{summary.marp_model_fit}
\title{Summarise a model-specific renewal-process fit}
\usage{
\method{summary}{marp_model_fit}(object, ...)
}
\arguments{
\item{object}{An object returned by one of the model-specific fitting
functions.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A structured object of class `summary_marp_model_fit`.
}
\description{
Summarise a model-specific renewal-process fit
}
