% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperT.R
\name{upperT}
\alias{upperT}
\title{Evaluate the upper-limit equation for a model-averaged T statistic}
\usage{
upperT(up, hat, sigmasq, Tstar, weights, B, alpha)
}
\arguments{
\item{up}{Candidate upper confidence limit.}

\item{hat}{Vector of model-specific estimates.}

\item{sigmasq}{Vector of model-specific variance estimates.}

\item{Tstar}{Matrix of bootstrap T statistics, with models in rows.}

\item{weights}{Vector of model weights.}

\item{B}{Number of bootstrap samples represented in `Tstar`.}

\item{alpha}{Significance level.}
}
\value{
A numeric scalar giving the upper-limit root equation value.
}
\description{
Evaluate the upper-limit equation for a model-averaged T statistic
}
\examples{
# set some parameters
up <- 100 # upper bound
hat <- rep(150, 6) # estimates obtained from each model
sigmasq <- 10 # variance
Tstar <- matrix(rep(100,600),6,100) # T statistics estimated from bootstrap samples
weights <- rep(1/6, 6) # model weights
B <- 100 # number of bootstrapped samples
alpha <- 0.05 # confidence level

# calculate the upper limit of T statistics
res <-  marp::upperT(up, hat, sigmasq, Tstar, weights, B, alpha)

res

}
