% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt_logl.R
\name{bpt_logl}
\alias{bpt_logl}
\title{A function to calculate the log-likelihood of BPT model}
\usage{
bpt_logl(param, x)
}
\arguments{
\item{param}{parameters of BPT model}

\item{x}{input data for BPT model}
}
\value{
returns the value of negative log-likelihood of the BPT model
}
\description{
A function to calculate the log-likelihood of BPT model
}
