% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dllog.R
\name{dllog}
\alias{dllog}
\title{Density function of Log-Logistics model}
\usage{
dllog(x, shape = 1, scale = 1, log = FALSE)
}
\arguments{
\item{x}{input data for Log-Logistics model}

\item{shape}{shape parameter of Log-Logistics model}

\item{scale}{scale parameter of Log-Logistics model}

\item{log}{logic function to determine whether log of logistics to be returned}
}
\value{
returns the denisty of the Log-Logistics model
}
\description{
Density function of Log-Logistics model
}
