% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_logl.R
\name{gamma_logl}
\alias{gamma_logl}
\title{A function to calculate the log-likelihood of Gamma model}
\usage{
gamma_logl(param, x)
}
\arguments{
\item{param}{parameters of Gamma model}

\item{x}{input data for Gamma model}
}
\value{
returns the value of negative log-likelihood of the Gamma model
}
\description{
A function to calculate the log-likelihood of Gamma model
}
