% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_rp.R
\name{gamma_rp}
\alias{gamma_rp}
\title{A function to fit Gamma renewal model}
\usage{
gamma_rp(data, t, m, y)
}
\arguments{
\item{data}{A numeric vector of positive inter-event times.}

\item{t}{A numeric vector of time points at which log-hazards are evaluated.}

\item{m}{A positive integer controlling repeated random-start optimizations;
the current implementation seeks `m - 1` acceptable `nlm()` fits.}

\item{y}{A time point at which the logit-transformed cumulative event
probability is evaluated.}
}
\value{
An object of class `marp_model_fit` retaining the following eight
  named list components:

\describe{
\item{par1}{Estimated shape parameter of the Gamma model}
\item{par2}{Estimated rate parameter of the Gamma model}
\item{logL}{Maximized log-likelihood}
\item{AIC}{Akaike information criterion (AIC)}
\item{BIC}{Bayesian information criterion (BIC)}
\item{mu_hat}{Estimated mean inter-event time}
\item{pr_hat}{Logit-transformed cumulative event probability at `y`}
\item{haz_hat}{Log-hazard values at `t`}
}
}
\description{
A function to fit Gamma renewal model
}
\examples{
set.seed(42)
data <-  rgamma(100,3,0.01)

# set some parameters
m = 10  # repeated random-start optimization setting
t = seq(100, 200, by=10)  # time intervals
y = 304  # cut-off year for estimating probablity

# fit Gamma renewal model
fit <- marp::gamma_rp(data, t, m, y)
fit
summary(fit)

}
