% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_rp.R
\name{gamma_rp}
\alias{gamma_rp}
\title{A function to fit Gamma renewal model}
\usage{
gamma_rp(data, t, m, y)
}
\arguments{
\item{data}{input inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{m}{the number of iterations in nlm}

\item{y}{user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Gamma renewal model
}
\description{
A function to fit Gamma renewal model
}
