% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogis_logl.R
\name{loglogis_logl}
\alias{loglogis_logl}
\title{A function to calculate the log-likelihood of Log-Logistics model}
\usage{
loglogis_logl(param, x)
}
\arguments{
\item{param}{as parameters of Log-Logistics model}

\item{x}{as input data for Log-Logistics model}
}
\value{
returns the value of negative log-likelihood of the Log-Logistics model
}
\description{
A function to calculate the log-likelihood of Log-Logistics model
}
