% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogis_rp.R
\name{loglogis_rp}
\alias{loglogis_rp}
\title{A function to fit Log-Logistics renewal model}
\usage{
loglogis_rp(data, t, m, y)
}
\arguments{
\item{data}{as input inter-event times}

\item{t}{as user-specified time intervals (used to compute hazard rate)}

\item{m}{as the number of iterations in nlm}

\item{y}{as user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Log-Logistics renewal model
}
\description{
A function to fit Log-Logistics renewal model
}
