% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognorm_bstrp.R
\name{lognorm_bstrp}
\alias{lognorm_bstrp}
\title{A function to generate (double) bootstrap samples and fit Log-Normal renewal model}
\usage{
lognorm_bstrp(n, t, B, BB, par_hat, mu_hat, pr_hat, haz_hat, y)
}
\arguments{
\item{n}{as number of inter-event times}

\item{t}{as user-specified time intervals (used to compute hazard rate)}

\item{B}{as number of bootstrap samples}

\item{BB}{as number of double-bootstrap samples}

\item{par_hat}{as estimated parameters}

\item{mu_hat}{as estimated mean inter-event times}

\item{pr_hat}{as estimated time to event probability}

\item{haz_hat}{as estimated hazard rates}

\item{y}{as user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Log-Normal renewal model on (double) bootstarp samples
}
\description{
A function to generate (double) bootstrap samples and fit Log-Normal renewal model
}
