% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognorm_rp.R
\name{lognorm_rp}
\alias{lognorm_rp}
\title{A function to fit Log-Normal renewal model}
\usage{
lognorm_rp(data, t, y)
}
\arguments{
\item{data}{as input inter-event times}

\item{t}{as user-specified time intervals (used to compute hazard rate)}

\item{y}{as user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Log-Normal renewal model
}
\description{
A function to fit Log-Normal renewal model
}
