% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerT.R
\name{lowerT}
\alias{lowerT}
\title{An utility function to calculate lower limit of T statistics}
\usage{
lowerT(low, hat, sigmasq, Tstar, weights, B, alpha)
}
\arguments{
\item{low}{lower limit}

\item{hat}{estimates}

\item{sigmasq}{variance}

\item{Tstar}{T statistics estimated from bootstrap samples}

\item{weights}{model weights}

\item{B}{number of bootstraps}

\item{alpha}{significance level}
}
\value{
returns list of percentile bootstrap intervals (including the best and generating model approach).
}
\description{
An utility function to calculate lower limit of T statistics
}
