% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp.R
\name{marp}
\alias{marp}
\title{A function to apply model-averaged renewal process}
\usage{
marp(data, t, m, y, which.model = 1)
}
\arguments{
\item{data}{input inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{m}{the number of iterations in nlm}

\item{y}{user-specified time point (used to compute time-to-event probability)}

\item{which.model}{user-specified genearting (or true underlying if known) model}
}
\value{
returns list of estimates obtained from different renewal processes and after applying model-averaging
}
\description{
A function to apply model-averaged renewal process
}
\examples{
# load example dataset (generated with rgamma(100,3,0.01))
data_file <- system.file("extdata", "small.txt", package = "marp", mustWork = TRUE)
data <- read.table(data_file)$V1

# set some parameters
m = 10  # number of iterations for MLE optimization
t = seq(100, 200, by=10)  # time intervals
y = 304  # cut-off year for estimating probablity
model_gen = 2  # underlying true model

# model selection and averaging
result <- marp::marp(data, t, m, y, which.model = model_gen)

}
