% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp.R
\name{marp}
\alias{marp}
\title{A function to apply model-averaged renewal process}
\usage{
marp(data, t, m, y, which.model = 1)
}
\arguments{
\item{data}{input inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{m}{the number of iterations in nlm}

\item{y}{user-specified time point (used to compute time-to-event probability)}

\item{which.model}{user-specified genearting (or true underlying if known) model}
}
\value{
returns list of estimates obtained from different renewal processes and after applying model-averaging
}
\description{
A function to apply model-averaged renewal process
}
