% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_bstrp.R
\name{marp_bstrp}
\alias{marp_bstrp}
\title{A function to fit model-averaged renewal process}
\usage{
marp_bstrp(n, t, B, BB, m, par_hat, mu_hat, pr_hat, haz_hat, y)
}
\arguments{
\item{n}{number of inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{B}{number of bootstrap samples}

\item{BB}{number of double-bootstrap samples}

\item{m}{the number of iterations in nlm}

\item{par_hat}{estimated parameters}

\item{mu_hat}{estimated mean inter-event times}

\item{pr_hat}{estimated time to event probability}

\item{haz_hat}{estimated hazard rates}

\item{y}{user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting different renewal models on (double) bootstarp samples
}
\description{
A function to fit model-averaged renewal process
}
