% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marp_confint.R
\name{marp_confint}
\alias{marp_confint}
\title{A function to apply model-averaged renewal process}
\usage{
marp_confint(data, m, t, B, BB, alpha, y, which.model)
}
\arguments{
\item{data}{input inter-event times}

\item{m}{the number of iterations in nlm}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{B}{number of bootstrap samples}

\item{BB}{number of double-bootstrap samples}

\item{alpha}{signifiance level}

\item{y}{user-specified time point (used to compute time-to-event probability)}

\item{which.model}{user-specified genearting (or true underlying if known) model}
}
\value{
returns list of estimates obtained from different renewal processes and after applying model-averaging
}
\description{
A function to apply model-averaged renewal process
}
