% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_confint.R
\name{percent_confint}
\alias{percent_confint}
\title{A function to calculate percentile bootstrap confidence interval}
\usage{
percent_confint(data, B, t, m, y, which.model = 1)
}
\arguments{
\item{data}{input inter-event times}

\item{B}{number of bootstrap samples}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{m}{the number of iterations in nlm}

\item{y}{user-specified time point (used to compute time-to-event probability)}

\item{which.model}{user-specified genearting (or true underlying if known) model}
}
\value{
returns list of percentile bootstrap intervals (including the model-averaged approach).
}
\description{
A function to calculate percentile bootstrap confidence interval
}
