% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pllog.R
\name{pllog}
\alias{pllog}
\title{Propbability function of Log-Logistics model}
\usage{
pllog(q, shape = 1, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{input quantile for Log-Logistics model}

\item{shape}{shape parameter of Log-Logistics model}

\item{scale}{scale parameter of Log-Logistics model}

\item{lower.tail}{logic function to determine whether lower tail probability to be returned}

\item{log.p}{logic function to determine whether log of logistics to be returned}
}
\value{
returns the probability of the Log-Logistics model
}
\description{
Propbability function of Log-Logistics model
}
