% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_rp.R
\name{poisson_rp}
\alias{poisson_rp}
\title{A function to fit Poisson renewal model}
\usage{
poisson_rp(data, t, y)
}
\arguments{
\item{data}{input inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{y}{user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Poisson renewal model
\describe{
\item{par1}{Estimated parameter of the Poisson model}
\item{par2}{N/A, only keep it as a place holder for output formatting purpose}
\item{logL}{Negative log-likelihood}
\item{AIC}{Akaike information criterion (AIC)}
\item{BIC}{Bayesian information criterion (BIC)}
\item{mu_hat}{Estimated mean}
\item{pr_hat}{Estimated (logit) probabilities}
\item{haz_hat}{Estimated (log) hazard rates}
}
}
\description{
A function to fit Poisson renewal model
}
\examples{
set.seed(42)
data <-  rgamma(100,3,0.01)

# set some parameters
t = seq(100, 200, by=10)  # time intervals
y = 304  # cut-off year for estimating probablity

# fit Poisson renewal model
result <- marp::poisson_rp(data, t, y)

# print result
cat("par1 = ", result$par1, "\n")
cat("par2 = ", result$par2, "\n")
cat("logL = ", result$logL, "\n")
cat("AIC = ", result$AIC, "\n")
cat("BIC = ", result$BIC, "\n")
cat("mu_hat = ", result$mu_hat, "\n")
cat("pr_hat = ", result$pr_hat, "\n")

}
