% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_rp.R
\name{poisson_rp}
\alias{poisson_rp}
\title{A function to fit Poisson renewal model}
\usage{
poisson_rp(data, t, y)
}
\arguments{
\item{data}{A numeric vector of positive inter-event times.}

\item{t}{A numeric vector of time points at which log-hazards are evaluated.}

\item{y}{A time point at which the logit-transformed cumulative event
probability is evaluated.}
}
\value{
An object of class `marp_model_fit` retaining the following eight
  named list components:
\describe{
\item{par1}{Estimated exponential rate parameter of the Poisson renewal model}
\item{par2}{N/A, only keep it as a place holder for output formatting purpose}
\item{logL}{Maximized log-likelihood}
\item{AIC}{Akaike information criterion (AIC)}
\item{BIC}{Bayesian information criterion (BIC)}
\item{mu_hat}{Estimated mean inter-event time}
\item{pr_hat}{Logit-transformed cumulative event probability at `y`}
\item{haz_hat}{Log-hazard values at `t`}
}
}
\description{
A function to fit Poisson renewal model
}
\examples{
set.seed(42)
data <-  rgamma(100,3,0.01)

# set some parameters
t = seq(100, 200, by=10)  # time intervals
y = 304  # cut-off year for estimating probablity

# fit Poisson renewal model
fit <- marp::poisson_rp(data, t, y)
fit
summary(fit)

}
