% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_rp.R
\name{poisson_rp}
\alias{poisson_rp}
\title{A function to fit Poisson renewal model}
\usage{
poisson_rp(data, t, y)
}
\arguments{
\item{data}{input inter-event times}

\item{t}{user-specified time intervals (used to compute hazard rate)}

\item{y}{user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Poisson renewal model
}
\description{
A function to fit Poisson renewal model
}
