% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_confint.R
\name{student_confint}
\alias{student_confint}
\title{A function to calculate Studentized bootstrap confidence interval}
\usage{
student_confint(
  n,
  B,
  t,
  m,
  BB,
  par_hat,
  mu_hat,
  pr_hat,
  haz_hat,
  weights,
  alpha,
  y,
  best.model,
  which.model = 1
)
}
\arguments{
\item{n}{as number of inter-event times}

\item{B}{as number of bootstrap samples}

\item{t}{as user-specified time intervals (used to compute hazard rate)}

\item{m}{as the number of iterations in nlm}

\item{BB}{as number of double-bootstrap samples}

\item{par_hat}{as estimated parameters}

\item{mu_hat}{as estimated mean inter-event times}

\item{pr_hat}{as estimated time to event probability}

\item{haz_hat}{as estimated hazard rates}

\item{weights}{model weights}

\item{alpha}{significance level}

\item{y}{as user-specified time point (used to compute time-to-event probability)}

\item{best.model}{best model based on information criterion (i.e. AIC)}

\item{which.model}{as user-specified genearting (or true underlying if known) model}
}
\value{
returns list of Studentized bootstrap intervals (including the model-averaged approach).
}
\description{
A function to calculate Studentized bootstrap confidence interval
}
