% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperT.R
\name{upperT}
\alias{upperT}
\title{An utility function to calculate upper limit of T statistics}
\usage{
upperT(up, hat, sigmasq, Tstar, weights, B, alpha)
}
\arguments{
\item{up}{lower limit}

\item{hat}{estimates}

\item{sigmasq}{variance}

\item{Tstar}{T statistics estimated from bootstrap samples}

\item{weights}{model weights}

\item{B}{number of bootstraps}

\item{alpha}{significance level}
}
\value{
returns list of percentile bootstrap intervals (including the best and generating model approach).
}
\description{
An utility function to calculate upper limit of T statistics
}
