% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_logl.R
\name{weibull_logl}
\alias{weibull_logl}
\title{A function to calculate the log-likelihood of Weibull model}
\usage{
weibull_logl(param, x)
}
\arguments{
\item{param}{as parameters of Weibull model}

\item{x}{as input data for Weibull model}
}
\value{
returns the value of negative log-likelihood of the Weibull model
}
\description{
A function to calculate the log-likelihood of Weibull model
}
