% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_rp.R
\name{weibull_rp}
\alias{weibull_rp}
\title{A function to fit Weibull renewal model #' @import weibull_logl}
\usage{
weibull_rp(data, t, m, y)
}
\arguments{
\item{data}{as input inter-event times}

\item{t}{as user-specified time intervals (used to compute hazard rate)}

\item{m}{as the number of iterations in nlm}

\item{y}{as user-specified time point (used to compute time-to-event probability)}
}
\value{
returns list of estimates after fitting Weibull renewal model
}
\description{
A function to fit Weibull renewal model #' @import weibull_logl
}
